requirements.txt
auto_mix_prep==0.2.0
einops==0.8.1
fvcore==0.1.5.post20221221
h5pickle==0.4.2
matplotlib==3.9.4
numpy==1.23.5
onnxruntime==1.19.2
opencv_python==4.11.0.86
opencv_python_headless==4.11.0.86
opencv-contrib-python==4.11.0.86
pandas==2.2.3
Pillow==9.4.0
PyYAML==6.0.2
scikit_video==1.1.11
scipy==1.13.1
scikit-image==0.24.0
tqdm==4.67.1
tk==0.1.0